+++ WebApplication/Controllers/ValuesController.cs
@@ -7,14 +7,24 @@ namespace WebApplication.Controllers
     [ApiController]
     public class ValuesController : ControllerBase
     {
+        /// <summary>
+        /// Retornar a URL de onde encontra-se o fonte no github.
+        /// </summary>
+        /// <returns>URL contendo o repositório do Github com o código.</returns>
         [HttpGet("showmethecode")]
         public ActionResult<string> Get()
         {
             return "https://github.com/GreicyMatias/Asp.NetCoreSample";
         }
 
+        /// <summary>
+        /// Faz um cálculo de juros compostos
+        /// </summary>
+        /// <param name="valorinicial">Valor inicial em decimal</param>
+        /// <param name="meses">Valor inteiro, represetando tempo em meses</param>
+        /// <returns>Resultado do calculo</returns>
         [HttpGet("calculajuros")]
-        public ActionResult<string> Get([FromQuery(Name = "valorinicial")] int initialValue, [FromQuery(Name = "meses")] int months)
+        public ActionResult<string> Get([FromQuery(Name = "valorinicial")] double initialValue, [FromQuery(Name = "meses")] int months)
         {
             return (initialValue * Math.Pow((1 + 0.01), months)).ToString("N2");
         }
